requirements.txt
fastapi
uvicorn
langchain-groq
langchain-mcp-adapters
langgraph
qdrant-client
sentence-transformers
pypdf
tavily-python
fastmcp
python-dotenv
pydantic
requests